storage_backend::storage
Struct Storage

pub struct Storage { /* private fields */ }

Storage is limited to single threaded access due to the use of RefCell for transaction management.

Implementations

impl Storage

pub fn new(config: &StorageConfig) -> Result<Storage, StorageError>

pub fn open(config: &StorageConfig) -> Result<Storage, StorageError>

pub fn restore_backup<P: AsRef<Path>>( &self, backup_path: &P, ) -> Result<(), StorageError>

pub fn backup<P: AsRef<Path>>(&self, backup_path: P) -> Result<(), StorageError>

pub fn delete_db_files(storage: Storage) -> Result<(), StorageError>

pub fn delete(&self, key: &str) -> Result<(), StorageError>

pub fn transactional_delete( &self, key: &str, transaction_id: Uuid, ) -> Result<(), StorageError>

pub fn write(&self, key: &str, value: &str) -> Result<(), StorageError>

pub fn transactional_write( &self, key: &str, value: &str, transaction_id: Uuid, ) -> Result<(), StorageError>

pub fn read(&self, key: &str) -> Result<Option<String>, StorageError>

pub fn is_empty(&self) -> bool

pub fn keys(&self) -> Result<Vec<String>, StorageError>

pub fn partial_compare_keys( &self, key: &str, ) -> Result<Vec<String>, StorageError>

pub fn partial_compare( &self, key: &str, ) -> Result<Vec<(String, String)>, StorageError>

pub fn has_key(&self, key: &str) -> Result<bool, StorageError>

pub fn begin_transaction(&self) -> Uuid

Safety

This method uses std::mem::transmute to extend the transaction’s lifetime to 'static, which is safe in this context because all transactions are stored in a RefCell within the Storage struct, and are only accessed from the same thread. Ensure that all transactions are properly committed or rolled back to avoid resource leaks.

pub fn commit_transaction( &self, transaction_id: Uuid, ) -> Result<(), StorageError>

pub fn rollback_transaction( &self, transaction_id: Uuid, ) -> Result<(), StorageError>